requirements/dev.txt
-r base.txt

blessings==1.6
bpython==0.15
codeclimate-test-reporter==0.1.2
curtsies==0.2.9
greenlet==0.4.10
jedi==0.9.0
py==1.4.31
pygments==2.1.3
pytest-cov==2.3.1
pytest-django==3.0.0
pytest-xdist==1.15
pytest==3.0.2
requests==2.11.1
six==1.10.0
wcwidth==0.1.7
wheel==0.29.0
factory_boy==2.8.1
